% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{build_filter}
\alias{build_filter}
\title{Build the section of the path (link) responsible with filtering the data}
\usage{
build_filter(
  dataTypeIds = NULL,
  dataTypeGroupIds = NULL,
  dataTypeGroupId2s = NULL,
  dataProcessIds = NULL,
  dataProcessTypeIds = NULL,
  dataSourceShortNames = NULL,
  dataSourceYears = NULL,
  startYear = NULL,
  endYear = NULL,
  AgeStart = NULL,
  AgeEnd = NULL,
  indicatorTypeIds = NULL,
  indicatorIds = NULL,
  componentIds = NULL,
  isComplete = NULL,
  isActive = NULL,
  locIds = NULL,
  ids = NULL,
  locAreaTypeIds = NULL,
  subGroupIds = NULL,
  shortNames = NULL,
  addDefault = NULL,
  includeDependencies = NULL,
  includeFormerCountries = NULL,
  includeDataIDs = NULL,
  includeUncertainty = NULL,
  isSubnational = NULL,
  years = NULL,
  ageUnit = NULL,
  verbose
)
}
\description{
For a description of what each argument represents, see
https://population.un.org/demodata-api/swagger/index.html
}
\keyword{internal}
