% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Methods.R
\name{check_values}
\alias{check_values}
\title{Utility functions for wrappers. Name and value checking, for example.}
\usage{
check_values(x, expected.values, categorical = FALSE)
}
\arguments{
\item{x}{vector of values, \code{factor},\code{character},\code{numeric},\code{integer}, etc.}

\item{expected.values}{set of valid values if \code{categorical == TRUE}, otherwise the valid value range.}

\item{categorical}{logical. Does the value represent categories? Default \code{FALSE} for magnitudes or similar.}
}
\description{
A generic checking utility. This will need various
presets of values for the various possible columns, and may have its own
wrappers for each.
}
\details{
make sure values in a column are as expected
}
\examples{
expected.values <- c(0,1)
# can be a check on probabilities, perhaps they're in logit and we dont' realize it?
x <- runif(10)
x <- log(x/(1-x))
#out <- try(check_values(x,c(0,1),categorical = FALSE))
# de facto unit test..
#stopifnot(class(out) == "try-error")

# factors are checked against characters here.
y <- as.factor(c("a","b","c"))
check_values(y,letters,categorical = TRUE)
}
\keyword{internal}
