% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingMethods.R
\name{controlOutputMsg}
\alias{controlOutputMsg}
\title{Print messages}
\usage{
controlOutputMsg(fn, C)
}
\arguments{
\item{fn}{Method to be called from DemoTools}
}
\description{
Print messages
}
\keyword{internal}
