% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Methods.R
\name{output_msg}
\alias{output_msg}
\title{Print messages}
\usage{
output_msg(fn, arg_names)
}
\arguments{
\item{fn}{Method to be called from DemoTools. Available alternatives: 
\code{"beers", "grabill", "sprague"}.}
}
\description{
Print messages
}
\keyword{internal}
