% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDSQLtools-package.R
\docType{package}
\name{DDSQLtools-package}
\alias{DDSQLtools-package}
\title{DDSQLtools: Wrappers for DemoTools functions to work with UNPD SQL internal data format}
\description{
DDSQLtools: Wrappers for DemoTools functions to work with UNPD SQL internal data format
}
\author{
\itemize{
\item {Marius D. Pascariu}
\item {Tim Riffe}
}

Maintainer: Tim Riffe  <tim.riffe@gmail.com>
}
