% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDSQLtools-data.R
\docType{data}
\name{DDSQLtools.data}
\alias{DDSQLtools.data}
\title{DDSQLtools test data - Egypt 2006 - Population by age and sex}
\format{
An object of class \code{list} of length 5.
}
\source{
The population counts were downloaded using the \code{get_recorddata()}
function. The life table was imported from a file called
\code{LTabr_error1.csv}.
}
\usage{
DDSQLtools.data
}
\description{
DDSQLtools test data - Egypt 2006 - Population by age and sex
}
\details{
Check \code{data-raw/AddDataInPackage.R} to see how the data was selected.
}
\keyword{datasets}
