% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeTableMethods.R
\name{do_lifetable}
\alias{do_lifetable}
\title{Wrapper for Life Table Methods}
\usage{
do_lifetable(X, verbose = TRUE, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of columns as input data. 
Different numbers of rows. UN format.
}
\description{
Wrapper for Life Table Methods
}
\examples{
mx <- DDSQLtools.data$Mx5
do_lifetable(X = mx)
}
\seealso{
\code{\link[DemoTools]{lt_abridged}}
}
