% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingMethods.R
\name{do_splitting}
\alias{do_splitting}
\title{Wrapper for Age-Splitting Methods}
\usage{
do_splitting(X, fn = c("beers", "grabill", "sprague"), verbose = TRUE,
  ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{fn}{Method to be called from DemoTools. Available alternatives: 
\code{"beers", "grabill", "sprague"}.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of columns as input data. 
Different numbers of rows. UN format.
}
\description{
Wrapper for Age-Splitting Methods
}
\examples{
# Example 1 --- Abridged data
P5 <- DDSQLtools.data$Pop5_Egypt_M_DB

W1 <- do_splitting(P5, fn = "beers") 
W2 <- do_splitting(P5, fn = "grabill")
W3 <- do_splitting(P5, fn = "sprague")

# Example 2 --- 1-year age groups   
P1 <- DDSQLtools.data$Pop1_Egypt_M_DB

V1 <- do_splitting(P1, fn = "beers") 
V2 <- do_splitting(P1, fn = "grabill") 
V3 <- do_splitting(P1, fn = "sprague") 

select_columns <- c("AgeID", "AgeStart", "AgeMid", "AgeEnd", "AgeLabel", 
                    "DataTypeName", "DataTypeID", "DataValue")
W1[, select_columns]
V1[, select_columns]
}
\seealso{
\code{\link[DemoTools]{beers}}, 
\code{\link[DemoTools]{grabill}}, 
\code{\link[DemoTools]{sprague}}.
}
