% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparisonMethods.R
\name{doCompare}
\alias{doCompare}
\title{Wrapper for Population Comparison}
\usage{
doCompare(pop1, pop2, fn = c("ID", "IRD"), verbose = TRUE, ...)
}
\arguments{
\item{pop1}{Input data for population 1. UN format.}

\item{pop2}{Input data for population 2. UN format.}

\item{fn}{Method to be called from DemoTools. Available aternatives: 
\code{"ID", "IRD"}.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrapper for Population Comparison
}
\examples{
# ----------------------------------------------
# 1-year age groups
P1m <- DDSQLtools.data$Pop1_Egypt_M_DB
P1f <- DDSQLtools.data$Pop1_Egypt_F_DB

C1 <- doCompare(pop1 = P1m, pop2 = P1f, fn = "ID")
C2 <- doCompare(pop1 = P1m, pop2 = P1f, fn = "IRD")

# ----------------------------------------------
# 5-year age groups
P5m <- DDSQLtools.data$Pop5_Egypt_M_DB
P5f <- DDSQLtools.data$Pop5_Egypt_F_DB

C3 <- doCompare(pop1 = P5m, pop2 = P5f, fn = "ID")
C4 <- doCompare(pop1 = P5m, pop2 = P5f, fn = "IRD")

select_columns <- c("AgeID", "AgeStart", "AgeMid", "AgeEnd", "AgeLabel",
                    "DataTypeName", "DataTypeID", "DataValue")
C <- rbind(C1, C2, C3, C4)
C[, select_columns]
}
\seealso{
\code{\link[DemoTools]{ID}}, 
\code{\link[DemoTools]{IRD}}.
}
