% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtrapolationMethods.R
\name{doExtrapolate}
\alias{doExtrapolate}
\title{Wrapper for Extrapolating Old-Age Human Mortality Data Using 
Various Mortality Models ("laws")}
\source{
The function is based on the methods implemented in the 
\code{MortalityLaws} R package.
}
\usage{
doExtrapolate(X, x_fit, x_extr, law = "kannisto", verbose = TRUE, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{x_fit}{Ages to be considered in estimating the mortality model parameters.
\code{x_fit} can be a subset of \code{x}. However, after the model is identifies
fitted values and residuals are computed for all ages in \code{x}.}

\item{x_extr}{Ages for which to extrapolate the death-rates.}

\item{law}{The name of the mortality law/model to be used.
The following options are available: \itemize{
  \item{\code{"kannisto"}} -- The Kannisto model;
  \item{\code{"kannisto_makeham"}} -- The Kannisto-Makeham model;
  \item{\code{"gompertz"}} -- The Gompertz model;
  \item{\code{"ggompertz"}} -- The Gamma-Gompertz model;
  \item{\code{"makeham"}} -- The Makeham model;
  \item{\code{"beard"}} -- The Beard model;
  \item{\code{"beard_makeham"}} -- The Beard-Makeham model;
  \item{\code{"quadratic"}} -- The Quadratic model.
  }}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to the
\code{\link[MortalityLaws]{MortalityLaw}} function.}
}
\value{
A data.frame having the same number of columns as input data. 
Different numbers of rows. UN format.
}
\description{
Wrapper for Extrapolating Old-Age Human Mortality Data Using 
Various Mortality Models ("laws")
}
\examples{
# Example 1 - abridged data
# Extrapolate old-age mortality data
E <- doExtrapolate(X = DDSQLtools.data$Mx5,           # data - UN format 
                   x_fit = c(75, 80, 85, 90),         # ages to be used in fitting
                   x_extr = c(90, 95, 100, 105, 110), # ages for which the extrapolation is applied
                   law = "kannisto")                  # the model
E

# Build 2 life table: 1st using the original data and the 2nd using the extended data
LT1 <- doLifeTable(X = DDSQLtools.data$Mx5)
LT2 <- doLifeTable(X = E)

LT1[LT1$IndicatorID == "ex", c("AgeStart", "DataValue")]
LT2[LT2$IndicatorID == "ex", c("AgeStart", "DataValue")]

dim(LT1)
dim(LT2) # note the 2nd life table has few extra rows
}
\seealso{
\code{\link[DemoTools]{lt_rule_m_extrapolate}}
\code{\link[MortalityLaws]{MortalityLaw}}
}
