% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapingMethods.R
\name{doHeaping}
\alias{doHeaping}
\title{Wrap function for the age-heaping methods}
\usage{
doHeaping(X, ageMin = 10, ageMax = 90, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{ageMin}{integer. The lowest age included in calculations. Default 10.}

\item{ageMax}{integer. The upper age bound used for calculations. Default 90.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\description{
Wrap function for the age-heaping methods: 
\code{\link[DemoTools]{Whipple}},
\code{\link[DemoTools]{Myers}},
\code{\link[DemoTools]{Bachi}},
\code{\link[DemoTools]{CoaleLi}},
\code{\link[DemoTools]{Noumbissi}} and
\code{\link[DemoTools]{Spoorenberg}}.
}
\examples{
P1 <- DDSQLtools.data$Pop1_Egypt_DB
W <- doHeaping(P1)

W[, c("DataProcessType", "DataValue")]

}
