% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityMethods.R
\name{doQualityChecks}
\alias{doQualityChecks}
\title{Perform age-sex male-female data quality checks}
\usage{
doQualityChecks(xy, xx, ageMin = NULL, ageMax = NULL, ...)
}
\arguments{
\item{xy}{Input data for male population. UN format.}

\item{xx}{Input data for female population. UN format.}

\item{ageMin}{Integer. The lower age bound used for calculations. 
If \code{NULL}, the smallest age in input data is considered. 
Default: \code{NULL}.}

\item{ageMax}{Integer. The upper age bound used for calculations. 
#' If \code{NULL}, the largest age in input data is considered. 
Default: \code{NULL}.}

\item{...}{Other arguments to be passed to \code{\link[DemoTools]{ageSexAccuracy}}.}
}
\description{
Perform age-sex male-female data quality checks
}
\examples{
M5 <- DDSQLtools.data$Pop5_Egypt_M_DB
F5 <- DDSQLtools.data$Pop5_Egypt_F_DB

Q <- doQualityChecks(M5, F5)

Q[, c("DataProcessType", "DataValue")]
}
\seealso{
\code{\link[DemoTools]{sexRatioScore}}, 
\code{\link[DemoTools]{ageSexAccuracy}},
\code{\link[DemoTools]{ageSexAccuracyDasGupta}}.
}
