% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmoothingMethods.R
\name{doSmoothing}
\alias{doSmoothing}
\title{Wrapper for Smoothing Populations in 5-year Age Groups}
\usage{
doSmoothing(X, method = c("Carrier-Farrag", "KKN", "Arriaga",
  "United Nations", "Strong", "Zigzag"), ageMin = 10, ageMax = 65,
  young.tail = c("Original", "Arriaga", "Strong", NA),
  old.tail = young.tail, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrapper for Smoothing Populations in 5-year Age Groups
}
\examples{
P5 <- DDSQLtools.data$Pop5_Egypt_M_DB

M <- c("Carrier-Farrag", "KKN", "Arriaga",
       "United Nations", "Strong", "Zigzag")

S1 <- doSmoothing(P5, method = M[1])
S2 <- doSmoothing(P5, method = M[2])
S3 <- doSmoothing(P5, method = M[3])
S4 <- doSmoothing(P5, method = M[4])
S5 <- doSmoothing(P5, method = M[5])
S6 <- doSmoothing(P5, method = M[6])

S <- cbind(S1$DataValue, S2$DataValue, S3$DataValue, 
           S4$DataValue, S5$DataValue, S2$DataValue)
dimnames(S) <- list(Age = S1$AgeLabel, SmoothingMethod = M)

S 
}
\seealso{
\code{\link[DemoTools]{agesmth}}
}
