% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{extractData}
\alias{extractData}
\title{Request data records directly from their ID}
\usage{
extractData(ids, save = FALSE)
}
\arguments{
\item{ids}{a character vector of ID's that identify each data point}

\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}
}
\value{
A data frame with the same rows as the number of ID's containing
the data values for that specific ID.
}
\description{
Request data records directly from their ID
}
\examples{

# TODO: Add good example once the data filtering of all data available
# is ready.

\dontrun{
extractData("35444654")
}

}
