% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{format.numeric.colums}
\alias{format.numeric.colums}
\title{Format data from character to numeric}
\usage{
\method{format}{numeric.colums}(X, exceptions)
}
\arguments{
\item{X}{data.frame}
}
\description{
When a data is downloaded from web it is saved as a list or
data.frame with columns containing strings of information (character format).
This function reads the values and if it sees in these columns only numbers
will convert the column to class numeric.
}
\keyword{internal}
