% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Methods.R
\name{format_output}
\alias{format_output}
\title{Prepare output table in wrap functions}
\usage{
format_output(X, G)
}
\arguments{
\item{X}{Input data. UN format.}

\item{G}{The G object in wrap functions}
}
\description{
Prepare output table in wrap functions
}
\keyword{internal}
