% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_ages}
\alias{get_ages}
\title{Get information about available ages (AgeID)}
\usage{
get_ages(ageUnit = "Year", save_file = FALSE, ...)
}
\arguments{
\item{ageUnit}{The age unit to return. Either \code{"Year"} (default) or
\code{"Month"}. An unknown unit returns an empty data frame.}

\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Get information about available ages (AgeID)
}
\examples{
\dontrun{
get_ages()
get_ages(ageUnit = "Month")
}
}
