% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_datacatalog}
\alias{get_datacatalog}
\title{Gets all DataCatalog records (DataCatalogs)}
\usage{
get_datacatalog(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Gets all DataCatalog records (DataCatalogs)
}
\details{
For an improved output, we combine DataCatalog data
with locations (\code{\link{get_locations}}),
dataprocess (\code{\link{get_dataprocess}}),
dataprocesstype (\code{\link{get_dataprocesstype}}) together with some
transformations (convert dates to date columns, renaming columns, etc...).
}
\examples{
\dontrun{

dc <- get_datacatalog()
head(dc[, c("DataCatalogID", "DataProcessType", "LocName")])

# Filter using locIds and DataProcessTypeIds
dc <- get_datacatalog(locIds = 76, dataProcessTypeIds = 2)


t2 <- get_datacatalog(isSubnational = TRUE)
head(t2)

t3 <- get_datacatalog(isSubnational = FALSE)
head(t3)

}
}
