% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_iitypes}
\alias{get_iitypes}
\title{Get information about available indicators (IndicatorID) and indicatortypeids (IndicatorTypeId)}
\usage{
get_iitypes(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Get information about available indicators (IndicatorID) and indicatortypeids (IndicatorTypeId)
}
\examples{
\dontrun{
I <- get_iitypes()
I

only_itypes <- get_iitypes(indicatorTypeIds = 6)
only_itypes

only_iids <- get_iitypes(indicatorIds = 229)
only_iids

components <- get_iitypes(componentIds = 4)
components
}
}
