% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_locationtypes}
\alias{get_locationtypes}
\title{Get information about available location types (LocAreaTypeID)}
\usage{
get_locationtypes(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Get information about available location types (LocAreaTypeID)
}
\examples{

\dontrun{
# Check what subgroups are available for:
P <- get_locationtypes(
  indicatorTypeIds = 8, # Population by age and sex indicator;
  locIds = "egypt",
  isComplete = "abridged"
)
P
}
}
