% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_loctypes}
\alias{get_loctypes}
\title{Get information about available location types (LocTypeID)}
\usage{
get_loctypes(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Location types are the geographic levels (World, Region, Country, ...).
Not to be confused with \code{\link{get_locationtypes}}, which returns the
location area types (Whole area, Urban, Rural, ...).
}
\examples{
\dontrun{
get_loctypes()
}
}
