% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_recorddata}
\alias{get_recorddata}
\title{Download structuredDataRecords data from the UNPD portal}
\usage{
get_recorddata(
  save_file = FALSE,
  verbose = TRUE,
  collapse_id_name = FALSE,
  ...
)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{verbose}{Whether to print the translated query from strings to digits
for faster queries. By default set to TRUE.}

\item{collapse_id_name}{A logical to whether collapse all ID-Name columns
into a single \code{haven_labelled} column. Default set to \code{FALSE}.
See details for more information.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Download structuredDataRecords data from the UNPD portal
}
\details{
\code{get_recorddata} directly queries the endpoint
\code{structuredDataRecords}. For a list of all options available for this
endpoint, see the parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html

\code{get_recorddata} allows the user to supply string names for all
arguments that have equivalent \code{get_*} functions. For example,
\code{get_iitypes} for \code{indicatorIds}. The string used for all
of these arguments should be the one from the column \code{Name} in
the response from the \code{get_*} functions.

By default, all results exclude columns related to the uncertainty
of the values requested (StandardErrorValue, ConfidenceInterval, etc...).
By setting the argument \code{includeUncertainty = TRUE}, the uncertainty
fields will be included in the final data frame.

Once the data is read from the API, some transformations are applied. If
\code{collapse_id_name} is set to \code{TRUE}, all of these columns are
merged with their exact ID pairs (such that only the Name columns
are kept and their ID pair are assigned as labels):

\itemize{
\item{Columns \code{AreaName}, \code{DataReliabilityName}, \code{SubGroupName}, \code{DataStatusName}, \code{DataTypeName}, \code{DataTypeGroupName}, \code{IndicatorName}, \code{LocName}, \code{LocAreaTypeName}, \code{LocTypeName}, \code{ModelPatternName}, \code{ModelPatternFamilyName}, \code{PeriodGroupName}, \code{PeriodTypeName}, \code{RegName}, \code{SexName}, \code{StatisticalConceptName}, \code{SubGroupTypeName} are converted to labelled factors with \code{\link[haven]{labelled}}}
}

For an example to understand better how to exploit this feature, see
the vignette section 'Identifying ID and Names of values'.

Regardless of the \code{collapse_id_name} value, these transformations are applied:

\itemize{
\item{\code{TimeStart} and \code{TimeEnd} are returned with format \code{'DD/MM/YYYY'}}
}
}
\examples{
\dontrun{

#  You can provide all strings, all codes, or a combination of both
Y <- get_recorddata(
  dataProcessTypeIds = "Census",
  indicatorTypeIds = 8, # and support numeric of string names
  locIds = "egypt", # all arguments are case insensitive
  locAreaTypeIds = "Whole area",
  subGroupIds = "Total or All groups",
  isComplete = "Abridged"
)

head(Y)

# Same thing only with codes
X <- get_recorddata(
  dataProcessTypeIds = 2, # Census
  indicatorTypeIds = 8, # Population by age and sex - abridged
  locIds = 818, # Egypt
  locAreaTypeIds = 2, # Whole area
  subGroupIds = 2, # Total or All groups
  isComplete = 0
) # Age Distribution: Abridged

head(X)

# Same thing but limited to DataSourceYears
X <- get_recorddata(
  dataProcessTypeIds = 2,
  indicatorTypeIds = 8,
  locIds = 818,
  locAreaTypeIds = 2,
  subGroupIds = 2,
  isComplete = 0,
  dataSourceYears = 1980
)

head(X)

# Same thing but limited to DataSourceShortNames and including the uncertainty
X <- get_recorddata(
  dataProcessTypeIds = 2,
  indicatorTypeIds = 8,
  locIds = 818,
  locAreaTypeIds = 2,
  subGroupIds = 2,
  isComplete = 0,
  dataSourceShortNames = "OECD 1980",
  includeUncertainty = TRUE
)

head(X)

# You can also request the Name-ID pairs of columns to be collapsed
# to only the Name columns and ID pair is assigned as a label
X <- get_recorddata(
  dataProcessTypeIds = 2,
  indicatorTypeIds = 8,
  locIds = 818,
  locAreaTypeIds = 2,
  subGroupIds = 2,
  isComplete = 0,
  dataSourceShortNames = "OECD 1980",
  includeUncertainty = TRUE,
  collapse_id_name = TRUE
)

# Some of the collapsed columns:
head(x[, c("IndicatorName", "LocTypeName")])
}

}
