% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_recorddataadditional}
\alias{get_recorddataadditional}
\title{Download structuredDataRecordsAdditional data from the UNPD portal}
\usage{
get_recorddataadditional(
  save_file = FALSE,
  verbose = TRUE,
  collapse_id_name = FALSE,
  ...
)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{verbose}{Whether to print the translated query from strings to digits
for faster queries. By default set to TRUE.}

\item{collapse_id_name}{A logical to whether collapse all ID-Name columns
into a single \code{haven_labelled} column. Default set to \code{FALSE}.
See details for more information.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Download structuredDataRecordsAdditional data from the UNPD portal
}
\details{
\code{get_recorddataadditional} directly queries the endpoint
\code{structuredDataRecordsadditional}. For a list of all options available for this
endpoint, see the parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html

\code{get_recorddataadditional} allows the user to supply string names for all
arguments that have equivalent \code{get_*} functions. For example,
\code{get_iitypes} for \code{indicatorIds}. The string used for all
of these arguments should be the one from the column \code{Name} in
the response from the \code{get_*} functions.

By default, all results exclude columns related to the uncertainty
of the values requested (StandardErrorValue, ConfidenceInterval, etc...).
By setting the argument \code{includeUncertainty = TRUE}, the uncertainty
fields will be included in the final data frame.

Once the data is read from the API, some transformations are applied. If
\code{collapse_id_name} is set to \code{TRUE}, all of these columns are
merged with their exact ID pairs (such that only the Name columns
are kept and their ID pair are assigned as labels):

\itemize{
\item{Columns \code{AreaName}, \code{DataReliabilityName}, \code{SubGroupName}, \code{DataStatusName}, \code{DataTypeName}, \code{DataTypeGroupName}, \code{IndicatorName}, \code{LocName}, \code{LocAreaTypeName}, \code{LocTypeName}, \code{ModelPatternName}, \code{ModelPatternFamilyName}, \code{PeriodGroupName}, \code{PeriodTypeName}, \code{RegName}, \code{SexName}, \code{StatisticalConceptName}, \code{SubGroupTypeName} are converted to labelled factors with \code{\link[haven]{labelled}}}
}

For an example to understand better how to exploit this feature, see
the vignette section 'Identifying ID and Names of values'.

Regardless of the \code{collapse_id_name} value, these transformations are applied:

\itemize{
\item{\code{TimeStart} and \code{TimeEnd} are returned with format \code{'DD/MM/YYYY'}}
}
}
\examples{
\dontrun{

# You can do pretty much all queries you would do with get_recorddata:
Y <- get_recorddata(
  dataProcessTypeIds = "Census",
  indicatorTypeIds = 8, # and support numeric of string names
  locIds = "egypt", # all arguments are case insensitive
  locAreaTypeIds = "Whole area",
  subGroupIds = "Total or All groups",
  isComplete = "Abridged"
)


# But there's additional parameters such as dataTypeGroupIds or
dataTypeGroupId2s:
dt <-
  get_recorddataadditional(
    dataTypeGroupIds = "Direct", # or 3
    indicatorTypeIds = 8,
    isComplete = 0,
    locIds = 818,
    locAreaTypeIds = 2,
    subGroupIds = 2
  )


dt <-
  get_recorddataadditional(
    dataTypeGroupId2s = "Population (sample tabulation)", # or 11
    indicatorTypeIds = 8,
    isComplete = 0,
    locIds = 818,
    locAreaTypeIds = 2,
    subGroupIds = 2
  )


# Also collapse Name-ID pairs
dt <-
  get_recorddataadditional(
    dataTypeGroupId2s = "Population (sample tabulation)", # or 11
    indicatorTypeIds = 8,
    isComplete = 0,
    locIds = 818,
    locAreaTypeIds = 2,
    subGroupIds = 2,
    collapse_name_id = TRUE
  )

head(dt[, c("IndicatorName", "LocTypeName")])
}
}
