% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getCountries}
\alias{getCountries}
\title{Download country names and LocID's}
\usage{
getCountries(addDefault = "false", includeDependencies = "false",
  includeFormerCountries = "false",
  server = "http://24.239.36.16:9654/un2/api/", ...)
}
\arguments{
\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}

\item{server}{The path to the database. Default: 
"http://24.239.36.16:9654/un2/api/"}

\item{...}{Other conditions that might define the path to data.}
}
\description{
Download country names and LocID's
}
\examples{
getCountries()
}
