% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getDataProcessTypes}
\alias{getDataProcessTypes}
\title{Download data-types and DataProcessTypeID's}
\usage{
getDataProcessTypes(server = "http://24.239.36.16:9654/un2/api/")
}
\arguments{
\item{server}{The path to the database. Default: 
"http://24.239.36.16:9654/un2/api/"}
}
\description{
Download data-types and DataProcessTypeID's
}
\examples{
getDataProcessTypes()
}
