% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getIndicators}
\alias{getIndicators}
\title{Download indicator names and IndicatorTypeID's}
\usage{
getIndicators(addDefault = "false",
  server = "http://24.239.36.16:9654/un2/api/", ...)
}
\arguments{
\item{addDefault}{Logical. Default: FALSE;}

\item{server}{The path to the database. Default: 
"http://24.239.36.16:9654/un2/api/"}

\item{...}{Other conditions that might define the path to data.}
}
\description{
Download indicator names and IndicatorTypeID's
}
\examples{
Ind <- getIndicators()
Ind[, c("IndicatorTypeID", "Name", "ShortName")]
}
