% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{getLocations}
\alias{getLocations}
\title{Get information about available locations (LocID)}
\usage{
getLocations(addDefault = "false", includeDependencies = "false",
  includeFormerCountries = "false",
  server = "http://24.239.36.16:9654/un2/api/", save = FALSE, ...)
}
\arguments{
\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}

\item{server}{The path to the database. Default: 
\code{"http://24.239.36.16:9654/un2/api/"};}

\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}

\item{...}{Other conditions that might define the path to data.}
}
\description{
Get information about available locations (LocID)
}
\examples{
getLocations()
}
