% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{getSubGroup}
\alias{getSubGroup}
\title{Get information about available sub-group-types (SubGroupTypeID)}
\usage{
getSubGroup(indicatorType = 8, loc = 818, isComplete = 0,
  server = "http://24.239.36.16:9654/un2/api/", save = FALSE, ...)
}
\arguments{
\item{indicatorType}{Indicator type ID as defined by the UNPD. Run the
\code{\link{getIndicators}} function to see the available options;}

\item{loc}{Location ID as defined by the UNPD. Run the
\code{\link{getLocations}} function to see the available options;}

\item{isComplete}{Default: NULL;}

\item{server}{The path to the database. Default: 
\code{"http://24.239.36.16:9654/un2/api/"};}

\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}

\item{...}{Other conditions that might define the path to data.}
}
\description{
Get information about available sub-group-types (SubGroupTypeID)
}
\examples{
# Check what subgroups are available for:
S <- getSubGroup(indicatorType = 8,  # Population by age and sex indicator;
                 loc = 818)          # Egypt
S
}
