% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{getSubGroups}
\alias{getSubGroups}
\title{Get information about available sub-group-types (SubGroupTypeID)}
\usage{
getSubGroups(save = FALSE, ...)
}
\arguments{
\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}

\item{...}{Other arguments that might define the path to data. Handle with 
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcess}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{getDataProcessTypes}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorType}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{getIndicators}} function to see the available options;
  \item{\code{isComplete}} -- isComplete;
  \item{\code{loc}} -- Location ID as defined by the UNPD. Run the
  \code{\link{getLocations}} function to see the available options;
  \item{\code{locAreaType}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{getLocationTypes}} function to see the available 
  options;
  \item{\code{subGroup}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{getSubGroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
Get information about available sub-group-types (SubGroupTypeID)
}
\examples{
# Check what subgroups are available for:
S <- getSubGroups(indicatorType = 8,  # Population by age and sex indicator;
                  loc = 818,          # Egypt
                  isComplete = 0)
S
}
