% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapPop5.R
\name{isOAG}
\alias{isOAG}
\title{Check whether the input data contains an open age interval}
\usage{
isOAG(X)
}
\arguments{
\item{X}{Input data. UN format.}
}
\description{
Check whether the input data contains an open age interval
}
\keyword{internal}
