% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{linkGenerator}
\alias{linkGenerator}
\title{API Link Generator Function}
\usage{
linkGenerator(
  server = getOption("unpd_server", "https://population.un.org/demodata-api/api/"),
  type,
  verbose,
  ...
)
}
\arguments{
\item{server}{The path to the database. Check if the "unpd_server" option is
is set. If not, defaults to \code{"https://population.un.org/demodata-api/api/"}}

\item{type}{Type of data. Various options are available.}

\item{verbose}{Logical for whether to print the final translated call
to numeric arguments. If the translation of arguments take a lot of time,
it can have a substantial reduction in timing. This only makes sense if
the user provided values as strings which need translation (such as a
country name like 'Haiti' rather than its actually country code).}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
API Link Generator Function
}
\details{
The link generator is based on the structure of the database
created by Dennis Butler (in late 2018). To change the server used to make
the requests, set this at the beginning of your script:
options(unpd_server = "fill this out").

When requesting data from the structured data format (usually called from
\code{\link{get_recorddata}}), the columns \code{TimeStart} and \code{TimeEnd}
are returned with format \code{DD/MM/YYYY}, where \code{DD} are days, \code{MM}
are months and \code{YYYY} are years.
}
\examples{
\dontrun{
# Link to country list
L1 <- linkGenerator(
  type = "locations",
  includeDependencies = "false",
  includeFormerCountries = "false"
)
L1

# Link to location types (for Egypt)
# With strings rather than codes
L2 <- linkGenerator(
  type = "locAreaTypes",
  indicatorTypeIds = "Population by sex",
  locIds = "Egypt",
  isComplete = "Abridged"
)
L2

# Link to subgroup types (for Egypt)
L3 <- linkGenerator(
  type = "subGroups",
  indicatorTypeIds = 8,
  locIds = 818,
  isComplete = 0
)
L3

# Link to indicator list
L4 <- linkGenerator(
  type = "indicators"
)
L4

# Link to data process type list
L5 <- linkGenerator(type = "dataProcessTypes")
L5
}
}
\keyword{internal}
