% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_imports.R
\docType{data}
\name{P5_Egypt1976}
\alias{P5_Egypt1976}
\title{UN data - Egypt 1976 - Population by age and sex - abridged}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 79 columns.}
\source{
United Nations -- DemoData_Export_Pop5_Egypt+India_DB.xlsx
}
\usage{
P5_Egypt1976
}
\description{
UN data - Egypt 1976 - Population by age and sex - abridged
}
\keyword{datasets}
