% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{read_API}
\alias{read_API}
\title{Download data}
\usage{
read_API(type, save_file, verbose = FALSE, ...)
}
\arguments{
\item{type}{Type of data. Various options are available.}

\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{verbose}{Logical for whether to print the final translated call
to numeric arguments. If the translation of arguments take a lot of time,
it can have a substantial reduction in timing. This only makes sense if
the user provided values as strings which need translation (such as a
country name like 'Haiti' rather than its actually country code).}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://population.un.org/demodata-api/swagger/index.html}
}
\description{
Download data
}
\keyword{internal}
