% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{save_data_in_working_directory}
\alias{save_data_in_working_directory}
\title{Save downloaded data in a .Rdata file located in the working directory}
\usage{
save_data_in_working_directory(data, file_name)
}
\arguments{
\item{data}{The dataset to be saved;}

\item{file_name}{Name to be assined to the data.}
}
\description{
Save downloaded data in a .Rdata file located in the working directory
}
\keyword{internal}
