% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{save_in_working_dir}
\alias{save_in_working_dir}
\title{Save downloaded data in a .Rdata file located in the working directory}
\usage{
save_in_working_dir(data, file_name)
}
\arguments{
\item{data}{The dataset to be saved;}

\item{file_name}{Name to be assined to the data.}
}
\description{
Save downloaded data in a .Rdata file located in the working directory
}
\keyword{internal}
