% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityMethods.R
\name{validate_input}
\alias{validate_input}
\title{Internal function that validates the input in \code{do_qualitychecks} as
stop it if data is not alright}
\usage{
validate_input(z)
}
\arguments{
\item{z}{A list containing the input supplied in \code{do_qualitychecks}.}
}
\value{
Nothing. Just lets you pass through... or not.
}
\description{
Internal function that validates the input in \code{do_qualitychecks} as
stop it if data is not alright
}
\keyword{internal}
