% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityMethods.R
\name{validateInput}
\alias{validateInput}
\title{Internal function that validates the input in \code{doQualityChecks} as
stop it if data is not alright}
\usage{
validateInput(z)
}
\arguments{
\item{z}{A list containing the input supplied in \code{doQualityChecks}.}
}
\value{
Nothing. Just lets you pass through... or not.
}
\description{
Internal function that validates the input in \code{doQualityChecks} as
stop it if data is not alright
}
\keyword{internal}
