% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapPop5.R
\name{wrapPop5}
\alias{wrapPop5}
\title{Wrap function for beers, grabill and sprague functions}
\usage{
wrapPop5(X, fn = c("beers", "grabill", "sprague"), ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{fn}{Method to be called from DemoTools}

\item{...}{Other arguments to be passed on to the \code{\link[DemoTools]{beers}}, 
\code{\link[DemoTools]{grabill}} or \code{\link[DemoTools]{sprague}} functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrap function for \code{\link[DemoTools]{beers}}, 
\code{\link[DemoTools]{grabill}} and \code{\link[DemoTools]{sprague}}
}
\examples{
W1 <- wrapPop5(P5_Egypt1976, fn = "beers") 
W2 <- wrapPop5(P5_Egypt1976, fn = "grabill")
W3 <- wrapPop5(P5_Egypt1976, fn = "sprague") 
}
