% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{build_filter}
\alias{build_filter}
\title{Build the section of the path (link) responsible with filtering the data}
\usage{
build_filter(dataProcess = NULL, startYear = NULL, endYear = NULL,
  indicatorType = NULL, isComplete = NULL, loc = NULL,
  locAreaType = NULL, subGroup = NULL, addDefault = NULL,
  includeDependencies = NULL, includeFormerCountries = NULL)
}
\arguments{
\item{dataProcess}{Default: NULL;}

\item{startYear}{Default: NULL;}

\item{endYear}{Default: NULL;}

\item{indicatorType}{Default: NULL;}

\item{isComplete}{Default: NULL;}

\item{loc}{Default: NULL;}

\item{locAreaType}{Default: NULL;}

\item{subGroup}{Default: NULL;}

\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}
}
\description{
Build the section of the path (link) responsible with filtering the data
}
\keyword{internal}
