% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{build_filter}
\alias{build_filter}
\title{Build the section of the path (link) responsible with filtering the data
# TODO: Describe each of the parameters in detail.}
\usage{
build_filter(
  dataProcessIds = NULL,
  dataProcessTypeIds = NULL,
  startYear = NULL,
  endYear = NULL,
  AgeStart = NULL,
  AgeEnd = NULL,
  indicatorTypeIds = NULL,
  indicatorIds = NULL,
  componentIds = NULL,
  isComplete = NULL,
  isActive = NULL,
  locIds = NULL,
  ids = NULL,
  locAreaTypeIds = NULL,
  subGroupIds = NULL,
  addDefault = NULL,
  includeDependencies = NULL,
  includeFormerCountries = NULL,
  includeDataIDs = NULL,
  verbose
)
}
\arguments{
\item{dataProcessIds}{}

\item{dataProcessTypeIds}{Data process ID as defined by the UNPD. Run
\code{\link{get_dataprocess}} to see the available options}

\item{startYear}{Start year. Default: NULL;}

\item{endYear}{End year. Default: NULL;}

\item{AgeStart}{}

\item{AgeEnd}{}

\item{indicatorTypeIds}{Indicator type ID as defined by the UNPD. Run the
\code{\link{get_indicatortypes}} function to see the available options;}

\item{indicatorIds}{}

\item{componentIds}{}

\item{isComplete}{isComplete;}

\item{isActive}{}

\item{locIds}{Location ID as defined by the UNPD. Run
\code{\link{get_locations}} to see the available options;}

\item{ids}{}

\item{locAreaTypeIds}{Location area type ID as defined by the UNPD. Run
\code{\link{get_locations}} to see the available options.}

\item{subGroupIds}{SubGroup ID as defined by the UNPD. Run
\code{\link{get_subgroups}} to see the available options;}

\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}

\item{includeDataIDs}{}

\item{verbose}{whether to print the optimized code for get_recorddata}

\item{server}{The path to the database. Check if the "unpd_server" option is
is set. If not, defaults to \code{"http://24.239.36.16:9654/un2/api/"}}

\item{type}{Type of data. Various options are available.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcessTypeIds}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{get_dataprocess}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorTypeIds}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{get_indicatortypes}} function to see the available options;
  \item{\code{isComplete}} -- isComplete is set to `2` ('Total') by default on the API;
  \item{\code{locIds}} -- Location ID as defined by the UNPD. Run the
  \code{\link{get_locations}} function to see the available options;
  \item{\code{locAreaTypeIds}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{get_locationtypes}} function to see the available 
  options;
  \item{\code{subGroupIds}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{get_subgroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
Build the section of the path (link) responsible with filtering the data
# TODO: Describe each of the parameters in detail.
}
\keyword{internal}
