% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{build_filter}
\alias{build_filter}
\title{Build the section of the path (link) responsible with filtering the data}
\usage{
build_filter(
  DataProcessIDs = NULL,
  startYear = NULL,
  endYear = NULL,
  AgeStart = NULL,
  AgeEnd = NULL,
  IndicatorTypeIDs = NULL,
  isComplete = NULL,
  isActive = NULL,
  locIds = NULL,
  ids = NULL,
  locAreaTypeIds = NULL,
  subGroupIds = NULL,
  addDefault = NULL,
  includeDependencies = NULL,
  includeFormerCountries = NULL,
  includeDataIDs = NULL,
  verbose
)
}
\arguments{
\item{DataProcessIDs}{}

\item{startYear}{Start year. Default: NULL;}

\item{endYear}{End year. Default: NULL;}

\item{AgeStart}{}

\item{AgeEnd}{}

\item{IndicatorTypeIDs}{}

\item{isComplete}{isComplete;}

\item{isActive}{}

\item{locIds}{Location ID as defined by the UNPD. Run the
\code{\link{get_locations}} function to see the available options;}

\item{locAreaTypeIds}{Location area type ID as defined by the UNPD. Run the
\code{\link{get_locations}} function to see the available options;}

\item{subGroupIds}{}

\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}

\item{verbose}{whether to print the optimized code for get_recorddata}

\item{dataProcess}{Data process ID as defined by the UNPD. Run the
\code{\link{get_dataprocess}} function to see the available options;
TODOOOOOOOOOOOOO!!}

\item{indicatorType}{Indicator type ID as defined by the UNPD. Run the
\code{\link{get_indicators}} function to see the available options;}

\item{subGroup}{SubGroup ID as defined by the UNPD.
Run the \code{\link{get_subgroups}} function to see the available options;}
}
\description{
Build the section of the path (link) responsible with filtering the data
}
\keyword{internal}
