% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{build_filter}
\alias{build_filter}
\title{Build the section of the path (link) responsible with filtering the data}
\usage{
build_filter(dataProcess = NULL, startYear = NULL, endYear = NULL,
  indicatorType = NULL, isComplete = NULL, loc = NULL,
  locAreaType = NULL, subGroup = NULL, addDefault = NULL,
  includeDependencies = NULL, includeFormerCountries = NULL)
}
\arguments{
\item{dataProcess}{Data process ID as defined by the UNPD. Run the
\code{\link{getDataProcessTypes}} function to see the available options;}

\item{startYear}{Start year. Default: NULL;}

\item{endYear}{End year. Default: NULL;}

\item{indicatorType}{Indicator type ID as defined by the UNPD. Run the
\code{\link{getIndicators}} function to see the available options;}

\item{isComplete}{Default: NULL;}

\item{loc}{Location ID as defined by the UNPD. Run the
\code{\link{getLocations}} function to see the available options;}

\item{locAreaType}{Default: NULL;}

\item{subGroup}{SubGroup ID as defined by the UNPD. 
Run the \code{\link{getSubGroup}} function to see the available options;}

\item{addDefault}{Logical. Default: FALSE;}

\item{includeDependencies}{Logical. Default: FALSE;}

\item{includeFormerCountries}{Logical. Default: FALSE;}
}
\description{
Build the section of the path (link) responsible with filtering the data
}
\keyword{internal}
