% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{controlOutputMsg2}
\alias{controlOutputMsg2}
\title{Print messages}
\usage{
controlOutputMsg2(fn, arg_names)
}
\arguments{
\item{fn}{Method to be called from DemoTools. Available alternatives: 
\code{"beers", "grabill", "sprague"}.}
}
\description{
Print messages
}
\keyword{internal}
