% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDSQLtools-data.R
\docType{data}
\name{DDSQLtools.data}
\alias{DDSQLtools.data}
\title{DDSQLtools test data - Egypt 1976 - Population by age and sex}
\format{An object of class \code{list} of length 4.}
\source{
UN Excel files: 
\code{DemoData_Export_Pop1_Egypt_DB.xlsx} and 
\code{DemoData_Export_Pop5_Egypt_DB.xlsx}.
}
\usage{
DDSQLtools.data
}
\description{
DDSQLtools test data - Egypt 1976 - Population by age and sex
}
\details{
Check \code{dev/AddDataInPackage.R} to see how the data was selected.
}
\keyword{datasets}
