% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDSQLtools-package.R
\docType{package}
\name{DDSQLtools}
\alias{DDSQLtools}
\alias{DDSQLtools-package}
\title{DDSQLtools: Wrappers for DemoTools Functions to Work with UNPD SQL Internal Data Format}
\description{
Make the DemoTools package click into the UNPD workflow.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/timriffe/DDSQLtools}
  \item Report bugs at \url{https://github.com/timriffe/DDSQLtools/issues}
}

}
\author{
\strong{Maintainer}: Tim Riffe \email{tim.riffe@gmail.com} (\href{https://orcid.org/0000-0002-2673-4622}{ORCID})

Authors:
\itemize{
  \item Marius D. Pascariu (\href{https://orcid.org/0000-0002-2568-6489}{ORCID})
  \item Jorge Cimentada (\href{https://orcid.org/0000-0001-5594-1156}{ORCID})
}

}
