% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeapingMethods.R
\name{doHeaping}
\alias{doHeaping}
\title{Wrapper for Age-Heaping Methods}
\usage{
doHeaping(X, fn = c("Whipple", "Myers", "Bachi", "CoaleLi", "Noumbissi",
  "Spoorenberg", "ageRatioScore", "KannistoHeap", "Jdanov"),
  verbose = TRUE, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{fn}{Method to be called from DemoTools. Available alternatives: 
\code{"beers", "grabill", "sprague"}.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\description{
Wrapper for Age-Heaping Methods
}
\examples{
P1 <- DDSQLtools.data$Pop1_Egypt_M_DB

H1 <- doHeaping(P1, fn = "Whipple")
H2 <- doHeaping(P1, fn = "Myers")
H3 <- doHeaping(P1, fn = "Bachi")
H4 <- doHeaping(P1, fn = "CoaleLi")
H5 <- doHeaping(P1, fn = "Noumbissi")
H6 <- doHeaping(P1, fn = "Spoorenberg")
H7 <- doHeaping(P1, fn = "ageRatioScore")
H8 <- doHeaping(P1, fn = "KannistoHeap")
H9 <- doHeaping(P1, fn = "Jdanov")

H <- rbind(H1, H2, H3, H4, H5, H6, H7, H8, H9)
select_columns <- c("AgeID", "AgeStart", "AgeMid", "AgeEnd", "AgeLabel",
                    "DataTypeName", "DataTypeID", "DataValue")
H[, select_columns]

# Silence the function with verbose = FALSE
H1 <- doHeaping(P1, fn = "Whipple", verbose = FALSE)
# ... or by specifying all arguments
H1 <- doHeaping(P1, fn = "Whipple", ageMin = 10, ageMax = 90, digit = 1)
}
\seealso{
\code{\link[DemoTools]{Whipple}},
\code{\link[DemoTools]{Myers}},
\code{\link[DemoTools]{Bachi}},
\code{\link[DemoTools]{CoaleLi}},
\code{\link[DemoTools]{Noumbissi}},
\code{\link[DemoTools]{Spoorenberg}},
\code{\link[DemoTools]{ageRatioScore}},
\code{\link[DemoTools]{KannistoHeap}},
\code{\link[DemoTools]{Jdanov}}.
}
