% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QualityMethods.R
\name{doQualityChecks}
\alias{doQualityChecks}
\title{Wrapper for Performing Age-Sex Male-Female Data Quality Checks}
\usage{
doQualityChecks(XY, XX, fn = c("sexRatioScore", "ageSexAccuracy",
  "ageSexAccuracyDasGupta"), verbose = TRUE, ...)
}
\arguments{
\item{XY}{Input data for male population. UN format.}

\item{XX}{Input data for female population. UN format.}

\item{fn}{Method to be called from DemoTools. Available aternatives: 
\code{"sexRatioScore", "ageSexAccuracy", "ageSexAccuracyDasGupta"}.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrapper for Performing Age-Sex Male-Female Data Quality Checks
}
\examples{
M5 <- DDSQLtools.data$Pop5_Egypt_M_DB
F5 <- DDSQLtools.data$Pop5_Egypt_F_DB

Q1 <- doQualityChecks(M5, F5, fn = "sexRatioScore")
Q2 <- doQualityChecks(M5, F5, fn = "ageSexAccuracy")
Q3 <- doQualityChecks(M5, F5, fn = "ageSexAccuracyDasGupta")

Q <- rbind(Q1, Q2, Q3)
Q[, c("DataProcessType", "DataValue")]
}
\seealso{
\code{\link[DemoTools]{sexRatioScore}}, 
\code{\link[DemoTools]{ageSexAccuracy}},
\code{\link[DemoTools]{ageSexAccuracyDasGupta}}.
}
