% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingMethods.R
\name{doSplitting}
\alias{doSplitting}
\title{Wrapper for Age-Splitting Methods}
\usage{
doSplitting(X, fn = c("beers", "grabill", "sprague"), verbose = TRUE, ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{fn}{Method to be called from DemoTools. Available aternatives: 
\code{"beers", "grabill", "sprague"}.}

\item{verbose}{Logical value. If \code{TRUE} messages are printed 
as the method is applied. Set \code{verbose = FALSE} to silent the function.}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrapper for Age-Splitting Methods
}
\examples{
# Example 1 --- Abridged data
P5 <- DDSQLtools.data$Pop5_Egypt_M_DB

W1 <- doSplitting(P5, fn = "beers") 
W2 <- doSplitting(P5, fn = "grabill")
W3 <- doSplitting(P5, fn = "sprague")

# Example 2 --- 1-year age groups   
P1 <- DDSQLtools.data$Pop1_Egypt_M_DB

V1 <- doSplitting(P1, fn = "beers") 
V2 <- doSplitting(P1, fn = "grabill") 
V3 <- doSplitting(P1, fn = "sprague") 
}
\seealso{
\code{\link[DemoTools]{beers}}, 
\code{\link[DemoTools]{grabill}}, 
\code{\link[DemoTools]{sprague}}.
}
