% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplittingMethods.R
\name{doSplitting}
\alias{doSplitting}
\title{Wrap function for the age-splitting methods: beers, grabill and sprague}
\usage{
doSplitting(X, fn = c("beers", "grabill", "sprague"), ...)
}
\arguments{
\item{X}{Input data. UN format.}

\item{fn}{Method to be called from DemoTools}

\item{...}{Other arguments to be passed on to other methods and functions.}
}
\value{
A data.frame having the same number of colums as input data. Different 
numbers of rows. UN format.
}
\description{
Wrap function for \code{\link[DemoTools]{beers}}, 
\code{\link[DemoTools]{grabill}} and \code{\link[DemoTools]{sprague}}
}
\examples{
# Example 1 --- Abridged data
P5 <- DDSQLtools.data$Pop5_Egypt_M_DB

W1 <- doSplitting(P5, fn = "beers") 
W2 <- doSplitting(P5, fn = "grabill")
W3 <- doSplitting(P5, fn = "sprague")

# Example 2 --- 1-year age group   
P1 <- DDSQLtools.data$Pop1_Egypt_M_DB

V1 <- doSplitting(P1, fn = "beers") 
V2 <- doSplitting(P1, fn = "grabill") 
V3 <- doSplitting(P1, fn = "sprague") 
}
