% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_datacatalog}
\alias{get_datacatalog}
\title{Gets all DataCatalog records (DataCatalogs)}
\usage{
get_datacatalog(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at http://24.239.36.16:9654/un3/swagger/ui/index#!/StructuredData/StructuredData_GetStructuredDataRecords}
}
\description{
Gets all DataCatalog records (DataCatalogs)
}
\examples{
\dontrun{
dc <- get_datacatalog()
head(dc[, c("PK_DataCatalogID", "Name", "ShortName")])

# Filter using locIds and DataProcessTypeIds
dc <- get_datacatalog(locIds = 76, dataProcessTypeIds = 2)

}
}
