% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_dataprocess}
\alias{get_dataprocess}
\title{Get information about available data-types (DataProcessTypeID)}
\usage{
get_dataprocess(save = FALSE, ...)
}
\arguments{
\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcessIds}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{get_dataprocess}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorTypeIds}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{get_indicators}} function to see the available options;
  \item{\code{isComplete}} -- isComplete;
  \item{\code{locIds}} -- Location ID as defined by the UNPD. Run the
  \code{\link{get_locations}} function to see the available options;
  \item{\code{locAreaTypeIds}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{get_locationtypes}} function to see the available 
  options;
  \item{\code{subGroupIds}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{get_subgroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
Get information about available data-types (DataProcessTypeID)
}
\examples{
\dontrun{
D <- get_dataprocess()
D[, c("PK_DataProcessTypeID", "Name", "ShortName")]
}
}
