% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_recorddata}
\alias{get_recorddata}
\title{Download data from UNPD portal}
\usage{
get_recorddata(save_file = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}

\item{verbose}{Whether to print the translated query from strings to digits
for faster queries. By default set to TRUE.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcessIds}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{get_dataprocess}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorTypeIds}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{get_indicatortypes}} function to see the available options;
  \item{\code{isComplete}} -- isComplete is set to `2` ('Total') by default on the API;
  \item{\code{locIds}} -- Location ID as defined by the UNPD. Run the
  \code{\link{get_locations}} function to see the available options;
  \item{\code{locAreaTypeIds}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{get_locationtypes}} function to see the available 
  options;
  \item{\code{subGroupIds}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{get_subgroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
Download data from UNPD portal
}
\details{
Once the data is read from the API, some transformations are applied:

\itemize{
\item{Columns \code{AreaName}, \code{DataReliabilityName}, \code{SubGroupName}, \code{DataStatusName}, \code{DataTypeName}, \code{DataTypeGroupName}, \code{IndicatorName}, \code{LocName}, \code{LocAreaTypeName}, \code{LocTypeName}, \code{ModelPatternName}, \code{ModelPatternFamilyName}, \code{PeriodGroupName}, \code{PeriodTypeName}, \code{RegName}, \code{SexName}, \code{StatisticalConceptName}, \code{SubGroupTypeName} are convert to labelled factors with \code{\link[haven]{labelled}}}
\item{\code{TimeStart} and \code{TimeEnd} are returned with format \code{'DD/MM/YYYY'}}
}
}
\examples{

\dontrun{
#  You can provide all strings, all codes, or a combination of both
Y <- get_recorddata(dataProcessIds = "Census",
                   indicatorTypeIds = 8, # and support numeric of string names
                   locIds = "egypt", # all arguments are case insensitive
                   locAreaTypeIds = "Whole area",
                   subGroupIds = "Total or All groups",
                   isComplete = "Abridged")

head(Y)

# Same thing only with codes
X <- get_recorddata(dataProcessIds = 2,   # Census
                   indicatorTypeIds = 8, # Population by age and sex - abridged 
                   locIds = 818,         # Egypt
                   locAreaTypeIds = 2,   # Whole area 
                   subGroupIds = 2,      # Total or All groups
                   isComplete = 0)       # Age Distribution: Abridged

head(X)
}

}
