% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{get_seriesdata}
\alias{get_seriesdata}
\title{Get information about available details for a given series of data}
\usage{
get_seriesdata(save_file = FALSE, ...)
}
\arguments{
\item{save_file}{Logical. Choose whether or not to save the data in an
external \code{.Rdata} file in the working directory. Default: \code{FALSE}.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! For a list of all options available, see the
parameters for each endpoint at https://popdiv.dfs.un.org/Demodata/swagger/ui/index#/}
}
\description{
Get information about available details for a given series of data
}
\examples{
\dontrun{
# You can provide all strings, all codes, or a combination of both
G <- get_seriesdata(
  dataProcessTypeIds = 0:15, # possible processes
  indicatorTypeIds = 25, # M[x]
  locIds = "Egypt", # Egypt
  locAreaTypeIds = c("whole area", "rural", "urban"), # all possible types
  subGroupIds = 2
)
G
}
}
