% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInformation.R
\name{getDataProcessTypes}
\alias{getDataProcessTypes}
\title{Get information about available data-types (DataProcessTypeID)}
\usage{
getDataProcessTypes(server = "http://24.239.36.16:9654/un2/api/",
  save = FALSE)
}
\arguments{
\item{server}{The path to the database. Default: 
\code{"http://24.239.36.16:9654/un2/api/"};}

\item{save}{Logical. Choose whether or not to save the data in an external 
\code{.Rdata} file in the working directory. Default: 
\code{FALSE};}
}
\description{
Get information about available data-types (DataProcessTypeID)
}
\examples{
getDataProcessTypes()
}
