% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Methods.R
\name{is_OAG}
\alias{is_OAG}
\title{Check whether the input data contains an open age interval}
\usage{
is_OAG(X)
}
\arguments{
\item{X}{Input data. UN format.}
}
\value{
Logical.
}
\description{
Check whether the input data contains an open age interval
}
\examples{
p5 <- DDSQLtools.data$Pop5_Egypt_M_DB
is_OAG(p5)
}
