% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{linkGenerator}
\alias{linkGenerator}
\title{API Link Generator Function}
\usage{
linkGenerator(server = "http://24.239.36.16:9654/un3/api/", type, ...)
}
\arguments{
\item{server}{The path to the database. Default: 
\code{"http://24.239.36.16:9654/un2/api/"};}

\item{type}{Type of data. Various options are available.}

\item{...}{Other arguments that might define the path to data. All arguments
accept a numeric code which is interpreted as the code of the specific
product requested. Alternatively, you can supply the equivalent product
name as a string which is case insensitive (see examples). Handle with
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcessIds}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{getDataProcess}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorTypeIds}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{getIndicators}} function to see the available options;
  \item{\code{isComplete}} -- isComplete;
  \item{\code{locIds}} -- Location ID as defined by the UNPD. Run the
  \code{\link{getLocations}} function to see the available options;
  \item{\code{locAreaTypeIds}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{getLocationTypes}} function to see the available 
  options;
  \item{\code{subGroupIds}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{getSubGroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
API Link Generator Function
}
\details{
The link generator is based on the structure of the database 
created by Dennis Butler (in late 2018). If the web address or the structure 
of the database changes this will have to be updated as well.
}
\examples{
\dontrun{
# Link to country list
L1 <- linkGenerator(type = "locations",
                    addDefault = "false",
                    includeDependencies = "false",
                    includeFormerCountries = "false")
L1

# Link to location types (for Egypt)
# With strings rather than codes
L2 <- linkGenerator(type = "locAreaTypes",
                    indicatorTypeIds = "Population by sex",
                    locIds = "Egypt",
                    isComplete = "Abridged")
L2

# Link to subgroup types (for Egypt)
L3 <- linkGenerator(type = "subGroups",
                    indicatorTypeIds = 8,
                    locIds = 818,
                    isComplete = 0)
L3

# Link to indicator list
L4 <- linkGenerator(type = "indicators",
                    addDefault = "false")
L4

# Link to data process type list
L5 <- linkGenerator(type = "dataProcessTypes")
L5
}
}
