% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Gets.R
\name{linkGenerator}
\alias{linkGenerator}
\title{API Link Generator Function}
\usage{
linkGenerator(server = "http://24.239.36.16:9654/un2/api/",
  type = c("country", "locationType", "subGroup", "Indicator",
  "dataProcessTypes", "seriesDataDetail", "recordDataDetail"), ...)
}
\arguments{
\item{server}{The path to the database. Default: 
\code{"http://24.239.36.16:9654/un2/api/"};}

\item{type}{Type of data. Various options are available.}

\item{...}{Other arguments that might define the path to data. Handle with 
care, this is important! The following options are available: \itemize{
  \item{\code{dataProcess}} -- Data process ID as defined by the UNPD. 
  Run the \code{\link{getDataProcessTypes}} function to see the available 
  options;
  \item{\code{startYear}} -- Start year. Default: NULL;
  \item{\code{endYear}} -- End year. Default: NULL;
  \item{\code{indicatorType}} -- Indicator type ID as defined by the UNPD. 
  Run the \code{\link{getIndicators}} function to see the available options;
  \item{\code{isComplete}} -- isComplete;
  \item{\code{loc}} -- Location ID as defined by the UNPD. Run the
  \code{\link{getLocations}} function to see the available options;
  \item{\code{locAreaType}} -- Location area type ID as defined by the UNPD. 
  Run the \code{\link{getLocationTypes}} function to see the available 
  options;
  \item{\code{subGroup}} -- SubGroup ID as defined by the UNPD. Run the 
  \code{\link{getSubGroups}} function to see the available options;
  \item{\code{addDefault}} -- Logical. Default: FALSE;
  \item{\code{includeDependencies}} -- Logical. Default: FALSE;
  \item{\code{includeFormerCountries}} -- Logical. Default: FALSE.
  }}
}
\description{
API Link Generator Function
}
\details{
The link generator is based on the structure of the database 
created by Dennis Butler (in late 2018). If the web address or the structure 
of the database changes this will have to be updated as well.
}
\examples{
# Link to country list
L1 <- linkGenerator(type = "country",
                    addDefault = "false",
                    includeDependencies = "false",
                    includeFormerCountries = "false")
L1

# Link to location types (for Egypt)
L2 <- linkGenerator(type = "locationType",
                    indicatorType = 8,
                    loc = 818,
                    isComplete = 0)
L2

# Link to subgroup types (for Egypt)
L3 <- linkGenerator(type = "subGroup",
                    indicatorType = 8,
                    loc = 818,
                    isComplete = 0)
L3

# Link to indicator list
L4 <- linkGenerator(type = "Indicator",
                    addDefault = "false")
L4

# Link to data process type list
L5 <- linkGenerator(type = "dataProcessTypes")
L5
}
